I. Linked module parameters to .ini file.
II. Added greater functionality to manusYarpRatethread::run() so that the .CSV file only outputs columns for the values that you want to read from

III. Fixed the sample_rate param to link to the RATETHREAD call rate and took out the usleep() function as a result.

IV. Edited the README to add descriptions for the YARP version of the module.

V.  Other stuff such as things and such as stuff like that.

VI. Deleted unused lines of code and performed other general clean-up.